% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{add_header}
\alias{add_header}
\title{Generate a title header or sub-title header in html-format}
\usage{
add_header(text, size = 1, UNDERLINE = FALSE, EMPHASIZE = FALSE)
}
\arguments{
\item{text}{the text to convert to html-format}

\item{size}{a number between 1 (\code{htmltools::h1}) and 5 (
\code{htmltools::h5})}

\item{UNDERLINE}{logical; if \code{TRUE}, then \code{tags$u()} is executed on
the argument \code{text} which gives an underlined text}

\item{EMPHASIZE}{logical; if \code{TRUE}, then \code{tags$em()} is executed
on the argument \code{text} which gives an emphasized/italic text}
}
\value{
the corresponding html-format string
}
\description{
Allows for underlying of text and/or emphasizing and different header sizes.
Currently, standard header sizes i.e. \code{htmltools::h1} -
\code{htmltools::h5} are supported!
}
