% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{add_training_variables}
\alias{add_training_variables}
\title{Enhance Data with "Recoded" Training and Status Variables}
\usage{
add_training_variables(df)
}
\arguments{
\item{df}{A tibble containing the variables to be transformed.}
}
\value{
A tibble with transformed variables: training variables are recoded
  to ordered factors with levels indicating participation status or progress
  percentage. The `Fjernleder` variable is recoded as an unordered factor
  with levels for remote leadership status, and `Statusfullfoertkurs` is
  renamed and recoded to reflect course completion status with appropriate
  handling of missing values.
}
\description{
Applies a series of transformations to a tibble, focusing on recoding
specific training-related variables and adjusting other categorical variables
according to predefined conventions. It renames variables, recodes them as
ordered or unordered factors, and handles missing values appropriately.
}
