% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{break_vspace}
\alias{break_vspace}
\title{Generates vertical breaks of different sizes.}
\usage{
break_vspace(size)
}
\arguments{
\item{size}{a character of either 'small', 'medium', or 'large'}
}
\value{
a \code{htmltools::tagList} of \code{htmltools::tags$br()}, with the
  number of breaks specified in the Details section and depending on the
  argument to \code{size}
}
\description{
A 'small' sized break is defined as a single \code{htmltools::tags$br}, a
'medium' sized break are three \code{htmltools::tags$br}, and a 'large' break
are six \code{htmltools::tags$br}
}
