% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utils.R
\name{check_reactive_inputs}
\alias{check_reactive_inputs}
\title{Check if arguments to function call are reactive}
\usage{
check_reactive_inputs(..., type = "arg")
}
\arguments{
\item{...}{arbitrary number of arguments but must be of the same type as
specified in \code{type} below}

\item{type}{either of 'arg' or 'arg_list_elements', see Details on the
differences}
}
\value{
pure side effect function returning invisibly if no errors, otherwise
   stops
}
\description{
Helper function of two types (to be further extended to the mixed type):
\itemize{
   \item{\code{type == 'arg'} checks every element inside \code{...} directly
   via \code{shiny::is.reactive())}}
   \item{\code{type == 'arg_list_elements'} treets every element inside
   \code{...} as a list and check whether its (sub-)elements are reactive
   }
}
}
