% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_raw_SVV_2021}
\alias{data_raw_SVV_2021}
\title{Data of SVV from Norstat}
\format{
## `data_raw_SVV_2021`
A data frame with 1925 rows and 53 columns:
\describe{
  \item{SamGender}{the employeer gender}
  \item{....}{...}
}
}
\source{
as provided from Norstat in 2021
}
\usage{
data_raw_SVV_2021
}
\description{
Information on employees and their questionaire responses for 2021.
}
\keyword{datasets}
