% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_raw_SVV_2022}
\alias{data_raw_SVV_2022}
\title{Data of SVV from Norstat}
\format{
## `data_raw_SVV_2022`
A data frame with 1941 rows and 55 columns:
\describe{
  \item{SamGender}{the employeer gender}
  \item{....}{...}
}
}
\source{
as provided from Norstat in 2022
}
\usage{
data_raw_SVV_2022
}
\description{
Information on employees and their questionaire responses for 2022.
}
\keyword{datasets}
