% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_raw_SVV_2023}
\alias{data_raw_SVV_2023}
\title{Data of SVV from Norstat}
\format{
## `data_raw_SVV_2023`
A data frame with 2090 rows and 63 columns:
\describe{
  \item{SamGender}{the employeer gender}
  \item{....}{...}
}
}
\source{
as provided from Norstat in 2023
}
\usage{
data_raw_SVV_2023
}
\description{
Information on employees and their questionaire responses for 2023.
}
\keyword{datasets}
