% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{generate_new_names}
\alias{generate_new_names}
\title{Generate New Variable Names Based on Training Type}
\usage{
generate_new_names(name_var)
}
\arguments{
\item{name}{The original name of the variable.}
}
\value{
A character string representing the new variable name. If the
  original name contains "SumofCompleted", it is replaced with
  "T_[number]_c", indicating the training type and that it represents a
  count. If it contains "AverageofProgress", it is replaced with
  "T_[number]_perc_prog", indicating the training type and that it represents
  a percentage progress. Otherwise, the original name is returned.
}
\description{
This function generates new variable names by identifying the type of training
information contained within the variable name. It recognizes variables related
to the sum of completed training sessions and the average progress in training
sessions, renaming them according to a standardized naming convention.
}
