% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fct_global.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Returns the BK colour setting}
\usage{
get_color_palette()
}
\value{
invisble return; function is called for side effects only
}
\description{
Returns the BK colour setting
}
