% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{get_utdanning_c}
\alias{get_utdanning_c}
\title{Generate Utdanning_c Variable}
\usage{
get_utdanning_c(utd, lvl_range)
}
\arguments{
\item{utd}{The `utdanning` variable from the dataset, which may require
conversion to a factor if not already one. This variable represents
the educational levels of respondents.}

\item{lvl_range}{A vector specifying the levels to be considered "low"
(coded as 0). The complement set within the specified levels of `utd`
will be considered "high" (coded as 1). This parameter allows for
flexible adaptation to different datasets and educational level
classifications.}
}
\value{
A factor variable `utdanning_c` with levels "low" and "high",
        suitable for inclusion in the dataset for subsequent analysis. This
        recoded variable facilitates the comparison and analysis of data
        based on the binary categorization of educational attainment.
}
\description{
Creates the `utdanning_c` variable based on specified level ranges. This
function is designed to facilitate the recoding of educational levels into a
binary "low" and "high" categorization, enhancing the analytical utility of
the `utdanning` variable.
}
