% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{read_and_distinct}
\alias{read_and_distinct}
\title{Read and Distinct Data}
\usage{
read_and_distinct(path, file)
}
\arguments{
\item{path}{The path to the directory containing the data file.}

\item{file}{The filename of the data file to read.}
}
\value{
A tibble with distinct rows from the read .sav file.
}
\description{
Reads data from a .sav file and applies distinct operation to remove
duplicates.
}
