% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{recode_leder_c}
\alias{recode_leder_c}
\title{Recode Leder Variable}
\usage{
recode_leder_c(data, leder_var, method = "direct")
}
\arguments{
\item{data}{The dataset containing the leadership variable.}

\item{leder_var}{The leadership variable to be recoded. This should be the
name of the column in `data` that contains leadership information,
specified as a symbol or string.}

\item{method}{A character string specifying the recoding method. Valid
options are "direct" for direct factor assignment (used for 2022 and
2023 datasets) and "replace" for datasets requiring a replacement
logic before factor assignment (used for the 2021 dataset).}
}
\value{
The dataset with the `leder_c` variable recoded according to the
        specified method.
}
\description{
Recodes the `leder_c` variable based on the specified leadership variable and
dataset structure. Adapts to different recoding needs for different years.
}
