% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{recode_Q22}
\alias{recode_Q22}
\title{Recode Q22 Variable}
\usage{
recode_Q22(data)
}
\arguments{
\item{data}{The dataset containing the `Q22` variable.}
}
\value{
The dataset with the `Q22` variable recoded.
}
\description{
Recodes the `Q22` variable in the dataset, setting specific levels to a
uniform label. This function is designed to standardize responses across
different datasets.
}
