% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{recode_utdanning}
\alias{recode_utdanning}
\title{Recode Utdanning Variable}
\usage{
recode_utdanning(data, utdanning_name = NULL, levels_range)
}
\arguments{
\item{data}{The dataset containing the `utdanning` variable.}

\item{utdanning_name}{The name of the `utdanning` variable in the dataset to
be renamed, or NULL if no renaming is needed. Defaults to NULL.}

\item{levels_range}{A vector indicating the specific levels to be recoded.
This parameter allows for precise control over which levels are
considered "low" or "high".}
}
\value{
The dataset with the `utdanning` variable optionally renamed and
        recoded. This function ensures that the structure and integrity of
        the dataset are maintained while applying the necessary
        transformations.
}
\description{
Conditionally renames and recodes the `utdanning` variable based on specified
levels. Utilizes the `get_utdanning_c` function to directly generate
`utdanning_c`. If `utdanning_name` is not provided (NULL), no renaming is
performed.
}
