% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{recode_variable}
\alias{recode_variable}
\title{Recode Variable to Ordered Factor Based on Type}
\usage{
recode_variable(name_var, type)
}
\arguments{
\item{name_var}{The variable to be recoded, as a character vector.}

\item{type}{A character string indicating the type of recoding to apply:
either "completed" for sum of completed trainings or "progress" for average
progress in training sessions.}
}
\value{
An ordered factor with levels and labels corresponding to the
  specified type. For "completed" type, levels are "", "0", "1", with labels
  "ikke_deltatt", "paebegynt", "fullfoert". For "progress" type, levels
  include "none" and numeric progress levels, with labels "ikke_deltatt" and
  various "_perc" labels corresponding to the numeric progress levels.
}
\description{
Recodes a given variable to an ordered factor with levels and labels
appropriate to the type of training data it represents. It supports recoding
for both completed training counts and training progress percentages,
including handling special cases such as non-participation.
}
