% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSVVdcDB.R
\docType{package}
\name{TaskSVVdcDB}
\alias{TaskSVVdcDB}
\title{TaskSVVdcDB}
\description{
A package for XXXX
}
\section{Foo functions}{

The foo functions ...
}

\section{TO-DOS}{

\itemize{
\item improve documentation
\item write this package further
 }
}

