% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helper_data_raw_cleaning.R
\name{update_data_set}
\alias{update_data_set}
\title{Update and Save a Dataset to a Specified Path}
\usage{
update_data_set(data_set, data_set_name, pth = "./data/data_raw_SVV_XXX.rda")
}
\arguments{
\item{data_set}{The dataset to be updated and saved. This can be any R object
that you wish to save (e.g., data frame, list).}

\item{data_set_name}{A character string specifying the name under which the
dataset should be saved in the global environment. This name is also used
as the variable name inside the RDA file.}

\item{pth}{Optional. A character string specifying the path to the file where
the dataset should be saved. Defaults to "./data/data_raw_SVV_XXX.rda".}
}
\value{
Invisible. The function does not return a value but saves the
  specified dataset under the given name both in the global environment and
  to an RDA file at the specified path.
}
\description{
This function updates the global environment with a new or modified dataset
under a specified name and saves it to a specified path in RDA format. It is
useful for persisting changes to datasets during data processing workflows.
}
\details{
The function uses `assign` to update the global environment and
  `save` to write the dataset to an RDA file. It is important to ensure that
  the specified path exists and is writable. The function overwrites existing
  files without warning.
}
\examples{
\dontrun{
update_data_set(data_raw_SVV_2023_to_save, "data_raw_SVV_2023",
                "./data/data_raw_SVV_2023.rda")
}
}
